+++ yazi-core/src/mgr/commands/open.rs
@@ -1,20 +1,18 @@
-use std::{borrow::Cow, ffi::OsString};
+use std::borrow::Cow;
 
 use tracing::error;
-use yazi_boot::ARGS;
 use yazi_config::{YAZI, popup::PickCfg};
 use yazi_fs::File;
-use yazi_macro::emit;
 use yazi_plugin::isolate;
 use yazi_proxy::{MgrProxy, PickProxy, TasksProxy, options::OpenDoOpt};
-use yazi_shared::{MIME_DIR, event::{CmdCow, EventQuit}, url::Url};
+use yazi_shared::{MIME_DIR, event::CmdCow, url::Url};
 
 use crate::{mgr::Mgr, tab::Folder, tasks::Tasks};
 
 #[derive(Clone, Copy)]
-struct Opt {
-	interactive: bool,
-	hovered:     bool,
+pub(super) struct Opt {
+	pub(super) interactive: bool,
+	pub(super) hovered:     bool,
 }
 
 impl From<CmdCow> for Opt {
@@ -119,20 +117,5 @@ impl Mgr {
 			|| find(self.parent())
 			|| find(self.hovered_folder())
 			|| find(self.active().history.get(&p))
-	}
-
-	fn quit_with_selected(opt: Opt, selected: &[&Url]) -> bool {
-		if opt.interactive || ARGS.chooser_file.is_none() {
-			return false;
-		}
-
-		let paths = selected.iter().fold(OsString::new(), |mut s, &u| {
-			s.push(u.as_os_str());
-			s.push("\n");
-			s
-		});
-
-		emit!(Quit(EventQuit { selected: Some(paths), ..Default::default() }));
-		true
 	}
 }
